requirements.txt
amqp==5.3.1
aniso8601==7.0.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
Django==4.2.21
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-ip-geolocation==1.6.1
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
environ==1.0
Faker==35.2.2
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
mysqlclient==2.2.7
packaging==25.0
pillow==10.4.0
pkg_resources==0.0.0
promise==2.3
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
Rx==1.6.3
singledispatch==4.1.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
